T1 2026 - Final PR
