+++ src/app/(tabs)/statistics.jsx
@@ -530,4 +530,3 @@ const styles = StyleSheet.create({
     fontSize: 14,
   },
 });
-
